dashboard/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
httpx==0.25.2
python-multipart==0.0.6
aiofiles==23.2.1
google-generativeai>=0.5.0
openai>=1.3.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0